Inline struct accessors
This requires changing the encoding of fields (because of a Scala 3
possible bug I've yet not been able to reproduce) - syntactically this won't
have any impact
