+++ modules/bindgen/src/main/scala/render/struct.scala
@@ -3,236 +3,10 @@ package rendering
 
 def struct(struct: ResolvedStruct, line: Appender)(using
     c: Config,
+    ctx: Context,
     ar: AliasResolver
 ): Exported =
-  val rewriteRules = hack_recursive_structs(struct)
-  val structName = struct.name
-  val structType: CType.Struct = CType.Struct(
-    fields = struct.fields.map(_._2),
-    hints = Hints(struct.staticSize)
-  )
-  val rewrittenStructType: CType.Struct = CType.Struct(
-    fields = structType.fields.zipWithIndex.map: (tpe, idx) =>
-      rewriteRules.get(idx).map(_.newRawType).getOrElse(tpe),
-    hints = Hints(struct.staticSize)
-  )
-
-  val madeOpaque = c.rendering.matches(_.opaqueStruct)(structName.value)
-
-  val structIsOpaque =
-    if struct.fields.size > 22 then true
-    else if struct.fields.size == 0 then false
-    else
-      madeOpaque match
-        case None => false
-        case Some(value) =>
-          warning(
-            s"${structName.value} will be rendered as a static array, as requested by '$value' filter"
-          )
-          true
-
-  val finalStructType =
-    if structIsOpaque then structArrayType(rewrittenStructType)
-    else rewrittenStructType
-
-  def setter(name: String): String =
-    import Sanitation.*
-    sanitise(name) match
-      case Good | Escaped if structIsOpaque && name == "length" =>
-        "_length_="
-      case Good | Escaped => name + "_="
-      case Renamed(value) => value + "_="
-
-  def getter(name: String): String =
-    import Sanitation.*
-    sanitise(name) match
-      case Good | Escaped if structIsOpaque && name == "length" =>
-        warning(
-          s"Struct ${structName.value} has its `length` parameter renamed as `_length` because it's rewritten as CArray"
-        )
-        "_length"
-
-      case Good           => name
-      case Renamed(value) => value
-      case Escaped        => s"`$name`"
-    end match
-  end getter
-
-  if rewriteRules.nonEmpty then
-    trace(
-      s"Rewrite rules for struct ${struct.name}",
-      rewriteRules.toSeq
-        .sortBy(_._1)
-        .map(_._2)
-        .map(p => p.name.value -> p.newRawType)
-    )
-
-  renderComment(line, struct.meta)
-  line(s"opaque type $structName = ${scalaType(finalStructType)}")
-  line("")
-  objectBlock(line)(s"object ${sanitiseBeforeColon(structName.value)}") {
-    if struct.fields.nonEmpty then
-      val tag =
-        s"given _tag: Tag[$structName] = ${scalaTag(finalStructType)}"
-      line(tag)
-
-      line("")
-
-      line(
-        s"// Allocates ${structName} on the heap – fields are not initalised or zeroed out"
-      )
-      line(
-        s"def apply()(using Zone): Ptr[$structName] = scala.scalanative.unsafe.alloc[$structName](1)"
-      )
-
-      val namedFieldsWithIndex =
-        struct.fields.zipWithIndex.toVector
-      val namedFields = namedFieldsWithIndex.map(_._1)
-
-      val applyArgList = List.newBuilder[String]
-
-      namedFieldsWithIndex.map { case ((name, typ), idx) =>
-        val inputType =
-          rewriteRules.get(idx).map(_.newRichType).getOrElse(typ)
-        applyArgList.addOne(
-          s"${getter(name.value)} : ${scalaType(inputType)}"
-        )
-      }
-
-      val ignored = c.rendering.matches(_.noConstructor)(structName.value)
-
-      ignored match
-        case None =>
-          defBlock(line)(
-            s"def apply(${applyArgList.result.mkString(", ")})(using Zone): Ptr[$structName] ="
-          ) {
-            line(s"val ____ptr = apply()")
-            namedFields.filter(_._1.value.nonEmpty).foreach {
-              case (fieldName, _) =>
-                line(
-                  s"(!____ptr).${getter(fieldName.value)} = ${getter(fieldName.value)}"
-                )
-            }
-            line(s"____ptr")
-          }
-        case Some(v) =>
-          warning(
-            s"Not rendering the constructor for ${structName.value}, as requested by '$v' filter"
-          )
-      end match
-
-      line("")
-
-      defBlock(line)(s"extension (struct: $structName)") {
-        if !structIsOpaque then
-          namedFieldsWithIndex.filter(_._1._1.value.nonEmpty).foreach {
-            case ((fieldName, fieldType), idx) =>
-              val setterName = setter(fieldName.value)
-              val getterName = getter(fieldName.value)
-
-              rewriteRules.get(idx) match
-                case Some(rewrite) =>
-                  val typ = scalaType(rewrite.newRichType)
-                  line(
-                    s"def $getterName : $typ = struct._${idx + 1}.asInstanceOf[$typ]"
-                  )
-                  line(
-                    s"def $setterName(value: $typ): Unit = !struct.at${idx + 1} = value.asInstanceOf[${scalaType(rewrite.newRawType)}]"
-                  )
-
-                case None =>
-                  val typ = scalaType(fieldType)
-                  line(
-                    s"def $getterName : $typ = struct._${idx + 1}"
-                  )
-                  line(
-                    s"def $setterName(value: $typ): Unit = !struct.at${idx + 1} = value"
-                  )
-              end match
-          }
-        else
-          namedFieldsWithIndex.filter(_._1._1.value.nonEmpty).foreach {
-            case ((fieldName, fieldType), fieldOffset) =>
-              val typ = scalaType(fieldType)
-              val setterName = setter(fieldName.value)
-              val getterName = getter(fieldName.value)
-
-              line(
-                s"def $getterName: $typ = !struct.at(offsets($fieldOffset)).asInstanceOf[Ptr[$typ]]"
-              )
-              line(
-                s"def $setterName(value: $typ): Unit = !struct.at(offsets($fieldOffset)).asInstanceOf[Ptr[$typ]] = value"
-              )
-          }
-        end if
-
-      }
-      if structIsOpaque then
-        defBlock(line)(
-          start = s"val offsets: Array[Int] =",
-          defNameForEnd = Some("offsets")
-        ) {
-          line(s"val res = Array.ofDim[Int](${namedFieldsWithIndex.length})")
-
-          alignMethod.foreach(line(_))
-          line("")
-
-          def renderAlignment(tpe: CType): String =
-            lazy val defaultCase =
-              s"alignmentof[${scalaType(tpe)}].toInt"
-            tpe match
-              case CType.Reference(Name.Model(name, _)) =>
-                aliasResolver(name) match
-                  case u: CType.Union =>
-                    renderAlignment(u)
-                  case _ =>
-                    defaultCase
-
-              case CType.Arr(of, _) => renderAlignment(of)
-
-              case CType.Union(fields, _) if fields.nonEmpty =>
-                fields
-                  .map(t => s"alignmentof[${scalaType(t)}].toInt")
-                  .mkString("Array(", ", ", ").max")
-              case _ =>
-                defaultCase
-            end match
-          end renderAlignment
-
-          namedFieldsWithIndex.foreach { case ((_, fieldType), idx) =>
-            // val tpe = scalaType(fieldType)
-            if idx == 0 then
-              line(s"res(0) = align(0, ${renderAlignment(fieldType)})")
-            else
-              val prevTpe = scalaType(namedFieldsWithIndex(idx - 1)._1._2)
-              line(
-                s"res($idx) = align(res(${idx - 1}) + sizeof[$prevTpe].toInt, ${renderAlignment(fieldType)})"
-              )
-
-          }
-          line("res")
-        }
-      end if
-    else line(s"given _tag: Tag[$structName] = Tag.materializeCStruct0Tag")
-    end if
-
-    line("")
-  }
-
-  struct.anonymous.zipWithIndex.foreach { (anon, idx) =>
-    anon match
-      case s: ResolvedStruct =>
-        rendering.struct(s, line)
-      case u: ResolvedUnion =>
-        rendering.union(u, line)
-      case e: ResolvedEnum =>
-        rendering.enumeration(e, line)
-
-    if idx != struct.anonymous.length - 1 then line("")
-  }
-
-  Exported.Yes(structName.value)
-end struct
+  StructRenderer(struct, line).render()
 
 private val alignMethod =
   """
